- Created files : QuestionFactory, QuizQuestion, QuizResultsViewModel, QuizStepViewModel - Added folders : Models, Services - Refactored code in MovieQuizViewController
