+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,15 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Dassam on 12.04.2023.
+//
+
+import Foundation
+import UIKit
+
+struct QuizQuestion {
+    let image: String // строка с названием фильма
+    let text: String // строка с вопросом о рейтинге фильма
+    let correctAnswer: Bool // правильный ответ на вопрос
+}// модель каждого вопрса
