+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Dassam on 12.04.2023.
+//
+
+import Foundation
+import UIKit 
+
+struct QuizResultsViewModel {
+    let title: String // строка с заголовком алерта
+    let text: String // строка с текстом о количестве набранных очков
+    let buttonText: String // текст для кнопки алерта
+} //модель для финального алерта
