+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Dassam on 12.04.2023.
+//
+
+import Foundation
+import UIKit 
+
+struct QuizStepViewModel {
+    let image: UIImage // картинка с афишей фильма с типом UIImage
+    let question: String // вопрос о рейтинге квиза
+    let questionNumber: String // строка с порядковым номером этого вопроса
+} // модель для отображения вопроса на экране
