+++ MovieQuiz/Services/QuestionFactory.swift
@@ -1,0 +1,60 @@
+//
+//  QuestionFactory.swift
+//  MovieQuiz
+//
+//  Created by Dassam on 12.04.2023.
+//
+
+import Foundation
+
+class QuestionFactory: QuestionFactoryProtocol {
+    private let questions: [QuizQuestion] = [ // Мокап
+        QuizQuestion(
+            image: "The Godfather",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "The Dark Knight",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "Kill Bill",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "The Avengers",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "Deadpool",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "The Green Knight",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "Old",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: false),
+        QuizQuestion(
+            image: "The Ice Age Adventures of Buck Wild",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: false),
+        QuizQuestion(
+            image: "Tesla",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: false),
+        QuizQuestion(
+            image: "Vivarium",
+            text: "Рейтинг этого фильма \n больше чем 6?",
+            correctAnswer: false)
+    ] // Мок модель вопросов
+    
+    func requestNextQuestion() -> QuizQuestion? {                       // 1
+        guard let index = (0..<questions.count).randomElement() else {  // 2
+            return nil
+        }
+        return questions[safe: index]                           // 3
+    }
+}
